refactor: extract DailyCurationService from DailyCurationJob
